DisplayProducts view added,Database one-to-many relationship added between User and Product,AddNewProduct logic added

+++ ProductsDistribution/Models/ViewModels/ProductViewModel.cs
@@ -19,7 +19,7 @@ namespace ProductsDistribution.Models.ViewModels
 
         public double price { get; set; }
 
-        public float rating { get; set; }
+        public double rating { get; set; }
 
         public string category { get; set; }
     }
